Create OpenImage-GSI
To modify an image before installation, it can be mounted as a device on Linux machines.
To do this, give the script a name and put it in a folder together with the Android-GSI.img (e.g. lineage_xxx_.img). Then start the script with ...
# sudo bash script name
... start it.
Wait.
With the input "m" it will be mounted readable (root only).
Now changes can be made (as root).

For example replace the Goggel infested apps with free ones.
Look here: Kill Apps

Then again in the terminal
# sudo bash script name
And with "u" write the changes into the original image and unmount.
Have Fun! 

Translated with www.DeepL.com/Translator (free version)